AutoEvent 9.15.0 - The Among Us Update

+++ AutoEvent/Games/RedGreenLight/Plugin.cs
@@ -183,9 +183,10 @@ public class Plugin : Event<Config, Translation>, IEventMap
 
                 continue;
 
+            _countdown++;
+            if (player.IsGodModeEnabled) continue;
             Extensions.GrenadeSpawn(player.Position, 0.1f, 0.1f, 0);
             player.Kill(Translation.RedLose);
-            _countdown++;
         }
 
         if (_countdown > 0)
